Extract shared poll option voting row
